To execute node stop method in proper context
This gives context to the stop method, instead of calling it with a global this object. I am setting up a client in the start() method such as this.client = new Client(), using it in process by doing this.client.get(...) and then would like to do this.client.disconnect() in the stop method.